feat(gui): add extras endpoints to client
