+++ gui/src/components/ImageHistory.tsx
@@ -1,14 +1,14 @@
 import { doesExist, mustExist } from '@apextoaster/js-utils';
 import { Grid, Typography } from '@mui/material';
-import { useContext, ReactNode } from 'react';
+import { ReactNode, useContext } from 'react';
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useStore } from 'zustand';
 
 import { StateContext } from '../state.js';
+import { ErrorCard } from './card/ErrorCard.js';
 import { ImageCard } from './card/ImageCard.js';
 import { LoadingCard } from './card/LoadingCard.js';
-import { ErrorCard } from './card/ErrorCard.js';
 
 export function ImageHistory() {
   const history = useStore(mustExist(useContext(StateContext)), (state) => state.history);
