+++ gui/src/components/tab/Models.tsx
@@ -1,17 +1,20 @@
 import { mustExist } from '@apextoaster/js-utils';
 import { Accordion, AccordionDetails, AccordionSummary, Button, Stack } from '@mui/material';
-import * as React from 'react';
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import _ from 'lodash';
+import * as React from 'react';
 import { useStore } from 'zustand';
 
-import { StateContext } from '../../state.js';
-import { EditableList } from '../input/EditableList';
-import { DiffusionModelInput } from '../input/model/DiffusionModel.js';
+import { ClientContext, StateContext } from '../../state.js';
 import { SafetensorFormat } from '../../types.js';
+import { EditableList } from '../input/EditableList';
 import { CorrectionModelInput } from '../input/model/CorrectionModel.js';
-import { UpscalingModelInput } from '../input/model/UpscalingModel.js';
-import { ExtraSourceInput } from '../input/model/ExtraSource.js';
+import { DiffusionModelInput } from '../input/model/DiffusionModel.js';
 import { ExtraNetworkInput } from '../input/model/ExtraNetwork.js';
+import { ExtraSourceInput } from '../input/model/ExtraSource.js';
+import { UpscalingModelInput } from '../input/model/UpscalingModel.js';
+
+const { useContext } = React;
 // eslint-disable-next-line @typescript-eslint/unbound-method
 const { kebabCase }  = _;
 
@@ -20,6 +23,20 @@ export function Models() {
   const extras = useStore(state, (s) => s.extras);
   // eslint-disable-next-line @typescript-eslint/unbound-method
   const setExtras = useStore(state, (s) => s.setExtras);
+  const client = mustExist(useContext(ClientContext));
+
+  const serverExtras = useQuery(['extras'], async () => client.extras(), {
+    // staleTime: STALE_TIME,
+  });
+
+  async function writeExtras() {
+    const resp = await client.writeExtras(extras);
+  }
+
+  const query = useQueryClient();
+  const write = useMutation(writeExtras, {
+    onSuccess: () => query.invalidateQueries([ 'extras' ]),
+  });
 
   return <Stack spacing={2}>
     <Accordion>
@@ -130,6 +147,6 @@ export function Models() {
         />
       </AccordionDetails>
     </Accordion>
-    <Button color='warning'>Save & Convert</Button>
+    <Button color='warning' onClick={() => write.mutate()}>Save & Convert</Button>
   </Stack>;
 }
